fix imports for editing features of neonview, make sure dragselect is initialized before it's needed (when edit_mode button is created) and write initial draft of TextEditMode
