+++ src/TextView.js
@@ -30,9 +30,6 @@ class TextView {
     label.appendChild(input);
     block.prepend(label);
 
-    $('#edit_mode').on('click', () => {
-      this.setTextEdit();
-    });
     this.setTextViewControls();
     this.neonView.view.addUpdateCallback(this.updateTextViewVisibility.bind(this));
   }
@@ -45,42 +42,6 @@ class TextView {
     $('#displayText').on('click', () => {
       this.updateTextViewVisibility();
     });
-  }
-
-  /**
-  * set text to edit mode
-  */
-  setTextEdit () {
-    let spans = Array.from($('#syl_text').children('p').children('span'));
-    spans.forEach(span => {
-      $(span).off('click');
-      $(span).on('click', () => {
-        this.updateSylText(span);
-      });
-    });
-  }
-
-  /**
-  * Update the text for a single syl element
-  * @param {HTMLElement} span
-  */
-  updateSylText (span) {
-    let orig = formatRaw($(span).html());
-    let corrected = window.prompt('', orig);
-    if (corrected !== null && corrected !== orig) {
-      let editorAction = {
-        'action': 'setText',
-        'param': {
-          'elementId': $('#' + $(span).attr('class').replace('syl-select', '').trim()).attr('id'),
-          'text': corrected
-        }
-      };
-      this.neonView.edit(editorAction, this.neonView.view.getCurrentPage()).then((response) => {
-        if (response) {
-          this.neonView.updateForCurrentPage();
-        }
-      });
-    }
   }
 
   /**
@@ -155,16 +116,6 @@ class TextView {
     }
     return lyrics.replace(uniToDash, '-');
   }
-}
-
-/**
- * Format a string for prompting the user.
- * @param {string} rawString
- * @returns {string}
- */
-function formatRaw (rawString) {
-  let removeSymbol = /\u{25CA}/u;
-  return rawString.replace(removeSymbol, '').trim();
 }
 
 export { TextView as default };
